+++ tests/tests/operation_tests.cpp
@@ -635,7 +635,7 @@ BOOST_AUTO_TEST_CASE(comment_delete_apply)
         vote.voter = "alice";
         vote.author = "alice";
         vote.permlink = "test1";
-        vote.weight = SCORUM_100_PERCENT;
+        vote.weight = (int16_t)100;
         tx.operations.push_back(comment);
         tx.operations.push_back(vote);
         tx.set_expiration(db.head_block_time() + SCORUM_MIN_TRANSACTION_EXPIRATION_LIMIT);
@@ -655,7 +655,7 @@ BOOST_AUTO_TEST_CASE(comment_delete_apply)
         BOOST_TEST_MESSAGE("--- Test success deleting a comment with negative rshares");
 
         generate_block();
-        vote.weight = -1 * SCORUM_100_PERCENT;
+        vote.weight = -1 * (int16_t)100;
         tx.clear();
         tx.operations.push_back(vote);
         tx.operations.push_back(op);
@@ -3347,8 +3347,8 @@ BOOST_AUTO_TEST_CASE(decline_voting_rights_apply)
 
         ACTORS((alice)(bob));
         generate_block();
-        vest("alice", ASSET_SCR(10e+3));
-        vest("bob", ASSET_SCR(10e+3));
+        vest("alice", ASSET_SCR(100e+3));
+        vest("bob", ASSET_SCR(100e+3));
         generate_block();
 
         account_witness_proxy_operation proxy;
@@ -3381,7 +3381,7 @@ BOOST_AUTO_TEST_CASE(decline_voting_rights_apply)
         tx.operations.push_back(op);
         tx.set_expiration(db.head_block_time() + SCORUM_MAX_TIME_UNTIL_EXPIRATION);
         tx.sign(alice_private_key, db.get_chain_id());
-        SCORUM_REQUIRE_THROW(db.push_transaction(tx, 0), fc::exception);
+        SCORUM_REQUIRE_THROW(db.push_transaction(tx, 0), fc::assert_exception);
 
         BOOST_TEST_MESSAGE("--- successs cancelling a request");
         op.decline = false;
@@ -3432,7 +3432,7 @@ BOOST_AUTO_TEST_CASE(decline_voting_rights_apply)
         vote.voter = "alice";
         vote.author = "alice";
         vote.permlink = "test";
-        vote.weight = SCORUM_100_PERCENT;
+        vote.weight = (int16_t)100;
         tx.clear();
         tx.operations.push_back(comment);
         tx.operations.push_back(vote);
@@ -3461,13 +3461,13 @@ BOOST_AUTO_TEST_CASE(decline_voting_rights_apply)
         db.get<comment_vote_object, by_comment_voter>(
             boost::make_tuple(db.get_comment("alice", string("test")).id, db.get_account("alice").id));
 
-        vote.weight = 0;
+        vote.weight = (int16_t)0;
         tx.clear();
         tx.operations.push_back(vote);
         tx.sign(alice_private_key, db.get_chain_id());
         SCORUM_REQUIRE_THROW(db.push_transaction(tx, 0), fc::exception);
 
-        vote.weight = SCORUM_1_PERCENT * 50;
+        vote.weight = (int16_t)50;
         tx.clear();
         tx.operations.push_back(vote);
         tx.sign(alice_private_key, db.get_chain_id());
@@ -3551,6 +3551,8 @@ BOOST_AUTO_TEST_CASE(account_create_with_delegation_authorities)
 
         signed_transaction tx;
         ACTORS((alice));
+
+        vest("alice", ASSET_SCR(100e+3));
         generate_blocks(1);
         fund("alice", ASSET_SCR(1e+6));
 
@@ -3758,6 +3760,10 @@ BOOST_AUTO_TEST_CASE(delegate_vesting_shares_authorities)
         signed_transaction tx;
         ACTORS((alice)(bob))
 
+        vest("alice", ASSET_SCR(1e+6));
+
+        generate_block();
+
         const auto& account_service = db.account_service();
 
         const account_object& alice_vested = account_service.get_account("alice");
@@ -3814,6 +3820,7 @@ BOOST_AUTO_TEST_CASE(delegate_vesting_shares_apply)
         generate_block();
 
         vest("alice", ASSET_SCR(1e+9));
+        vest("bob", ASSET_SCR(100e+3));
 
         generate_block();
 
@@ -3885,7 +3892,7 @@ BOOST_AUTO_TEST_CASE(delegate_vesting_shares_apply)
         vote_op.voter = "bob";
         vote_op.author = "alice";
         vote_op.permlink = "foo";
-        vote_op.weight = SCORUM_100_PERCENT;
+        vote_op.weight = (int16_t)100;
         tx.set_expiration(db.head_block_time() + SCORUM_MAX_TIME_UNTIL_EXPIRATION);
         tx.operations.push_back(vote_op);
         tx.sign(bob_private_key, db.get_chain_id());
@@ -4126,153 +4133,6 @@ BOOST_AUTO_TEST_CASE(comment_beneficiaries_validate)
         op.extensions.clear();
         op.extensions.insert(b);
         op.validate();
-    }
-    FC_LOG_AND_RETHROW()
-}
-
-struct comment_benefactor_reward_visitor
-{
-    typedef void result_type;
-
-    database& _db;
-
-    std::map<account_name_type, asset> reward_map;
-
-    comment_benefactor_reward_visitor(database& db)
-        : _db(db)
-    {
-    }
-
-    void operator()(const comment_benefactor_reward_operation& op)
-    {
-        reward_map.insert(std::make_pair(op.benefactor, op.reward));
-    }
-
-    template <typename Op> void operator()(Op&&) const
-    {
-    } /// ignore all other ops
-};
-
-BOOST_AUTO_TEST_CASE(comment_beneficiaries_apply)
-{
-    try
-    {
-        BOOST_TEST_MESSAGE("Test Comment Beneficiaries");
-        ACTORS((alice)(bob)(sam)(dave))
-        generate_block();
-
-        comment_operation comment;
-        vote_operation vote;
-        comment_options_operation op;
-        comment_payout_beneficiaries b;
-        signed_transaction tx;
-
-        comment.author = "alice";
-        comment.permlink = "test";
-        comment.parent_permlink = "test";
-        comment.title = "test";
-        comment.body = "foobar";
-
-        tx.operations.push_back(comment);
-        tx.set_expiration(db.head_block_time() + SCORUM_MIN_TRANSACTION_EXPIRATION_LIMIT);
-        tx.sign(alice_private_key, db.get_chain_id());
-        db.push_transaction(tx);
-
-        BOOST_TEST_MESSAGE("--- Test failure on more than 8 benefactors");
-        b.beneficiaries.push_back(beneficiary_route_type(account_name_type("bob"), SCORUM_1_PERCENT));
-
-        for (size_t i = 0; i < 8; i++)
-        {
-            b.beneficiaries.push_back(beneficiary_route_type(
-                account_name_type(TEST_INIT_DELEGATE_NAME + fc::to_string(i)), SCORUM_1_PERCENT));
-        }
-
-        op.author = "alice";
-        op.permlink = "test";
-        op.allow_curation_rewards = false;
-        op.extensions.insert(b);
-        tx.clear();
-        tx.operations.push_back(op);
-        tx.sign(alice_private_key, db.get_chain_id());
-        SCORUM_REQUIRE_THROW(db.push_transaction(tx), chain::plugin_exception);
-
-        BOOST_TEST_MESSAGE("--- Test specifying a non-existent benefactor");
-        b.beneficiaries.clear();
-        b.beneficiaries.push_back(beneficiary_route_type(account_name_type("doug"), SCORUM_1_PERCENT));
-        op.extensions.clear();
-        op.extensions.insert(b);
-        tx.clear();
-        tx.operations.push_back(op);
-        tx.sign(alice_private_key, db.get_chain_id());
-        SCORUM_REQUIRE_THROW(db.push_transaction(tx), fc::assert_exception);
-
-        BOOST_TEST_MESSAGE("--- Test setting when comment has been voted on");
-        vote.author = "alice";
-        vote.permlink = "test";
-        vote.voter = "bob";
-        vote.weight = SCORUM_100_PERCENT;
-
-        b.beneficiaries.clear();
-        b.beneficiaries.push_back(beneficiary_route_type(account_name_type("bob"), 25 * SCORUM_1_PERCENT));
-        b.beneficiaries.push_back(beneficiary_route_type(account_name_type("sam"), 50 * SCORUM_1_PERCENT));
-        op.extensions.clear();
-        op.extensions.insert(b);
-
-        tx.clear();
-        tx.operations.push_back(vote);
-        tx.operations.push_back(op);
-        tx.sign(alice_private_key, db.get_chain_id());
-        tx.sign(bob_private_key, db.get_chain_id());
-        SCORUM_REQUIRE_THROW(db.push_transaction(tx), fc::assert_exception);
-
-        BOOST_TEST_MESSAGE("--- Test success");
-        tx.clear();
-        tx.operations.push_back(op);
-        tx.sign(alice_private_key, db.get_chain_id());
-        db.push_transaction(tx);
-
-        BOOST_TEST_MESSAGE("--- Test setting when there are already beneficiaries");
-        b.beneficiaries.clear();
-        b.beneficiaries.push_back(beneficiary_route_type(account_name_type("dave"), 25 * SCORUM_1_PERCENT));
-        op.extensions.clear();
-        op.extensions.insert(b);
-        tx.sign(alice_private_key, db.get_chain_id());
-        SCORUM_REQUIRE_THROW(db.push_transaction(tx), fc::assert_exception);
-
-        BOOST_TEST_MESSAGE("--- Payout and verify rewards were split properly");
-        tx.clear();
-        tx.operations.push_back(vote);
-        tx.sign(bob_private_key, db.get_chain_id());
-        db.push_transaction(tx, 0);
-
-        generate_blocks(db.get_comment("alice", string("test")).cashout_time - SCORUM_BLOCK_INTERVAL);
-
-        BOOST_REQUIRE_EQUAL(db.get_account("bob").balance, ASSET_SCR(0));
-        BOOST_REQUIRE_EQUAL(db.get_account("sam").balance, ASSET_SCR(0));
-
-        asset bob_vesting_before = db.get_account("bob").vesting_shares;
-        asset sam_vesting_before = db.get_account("sam").vesting_shares;
-
-        comment_benefactor_reward_visitor visitor(db);
-
-        db.post_apply_operation.connect([&](const operation_notification& note) { note.op.visit(visitor); });
-
-        generate_block();
-
-        validate_database();
-
-        BOOST_REQUIRE_EQUAL(visitor.reward_map.size(), size_t(2));
-
-        BOOST_REQUIRE(visitor.reward_map.find("bob") != visitor.reward_map.end());
-        BOOST_REQUIRE(visitor.reward_map.find("sam") != visitor.reward_map.end());
-
-        BOOST_REQUIRE_EQUAL(visitor.reward_map["bob"], (db.get_account("bob").vesting_shares - bob_vesting_before));
-        BOOST_REQUIRE_EQUAL(visitor.reward_map["sam"], (db.get_account("sam").vesting_shares - sam_vesting_before));
-
-        // clang-format off
-        BOOST_REQUIRE_EQUAL(db.get_comment("alice", string("test")).beneficiary_payout_value * db.get_dynamic_global_properties().get_vesting_share_price(),
-                            (visitor.reward_map["sam"] + visitor.reward_map["bob"]));
-        // clang-format on
     }
     FC_LOG_AND_RETHROW()
 }
